+++ doc/Changelog.md
@@ -2,7 +2,8 @@ Changelog {#Changelog}
 ============
 
 # git master (1.1.0)
-
+* [49](https://github.com/BlueBrain/Tide/pull/49):
+  Special handling of launcher window implemented in LoggingUtility
 * [38](https://github.com/BlueBrain/Tide/pull/38):
   Fix a segfault that occured when opening grayscale images.
 * [36](https://github.com/BlueBrain/Tide/pull/36):